Best time for a commit is when you've just finished breaking everything
Pulled commands into their own subfolder (these are our events)
Created an eventHandler class which takes an InputData object and pushes the relevant commands to an events queue.
The eventHandler adds and removes subscribing components based on what is the immediate area. This data is cached for fast gamedata manipulation via commands